added user-answer gate

+++ src/app/Http/Controllers/api/v1/thread/AnswerController.php
@@ -66,32 +66,31 @@ class AnswerController extends Controller
             'content' => 'required',
         ]);
 
-//        if (Gate::forUser(auth()->user())->allows('user-answer', $answer)) {
+        if (Gate::forUser(auth()->user())->allows('user-answer', $answer)) {
             resolve(AnswerRepository::class)->update($request,$answer);
-//
+
             return \response()->json([
                 'message' => 'answer updated successfully'
             ], Response::HTTP_OK);
-//        }
-//
-//        return \response()->json([
-//            'message' => 'access denied'
-//        ], Response::HTTP_FORBIDDEN);
+        }
+
+        return \response()->json([
+            'message' => 'access denied'
+        ], Response::HTTP_FORBIDDEN);
     }
 
     public function destroy(Answer $answer)
     {
-//        if (Gate::forUser(auth()->user())->allows('user-answer', $answer)) {
+        if (Gate::forUser(auth()->user())->allows('user-answer', $answer)) {
             resolve(AnswerRepository::class)->destroy($answer);
-
             return \response()->json([
                 'message' => 'answer deleted successfully'
             ], Response::HTTP_OK);
-//        }
+        }
 
-//        return \response()->json([
-//            'message' => 'access denied'
-//        ], Response::HTTP_FORBIDDEN);
+        return \response()->json([
+            'message' => 'access denied'
+        ], Response::HTTP_FORBIDDEN);
     }
 
 }
